Move getContentType to utils
